+++ baricharafriends/src/main/java/com/edw88/baricharafriends/data/SitiosRepository.kt
@@ -1,0 +1,7 @@
+package com.edw88.baricharafriends.data
+
+class SitiosRepository {
+
+    suspend fun getSitios() = ApiFactory.retrofit.getSitios()
+
+}
