+++ baricharafriends/src/main/java/com/edw88/baricharafriends/ui/drawer/DrawerActivity.kt
@@ -1,8 +1,6 @@
-package com.edw88.baricharafriends.drawer
+package com.edw88.baricharafriends.ui.drawer
 
 import android.os.Bundle
-import android.view.Menu
-import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.navigation.NavigationView
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
