+++ baricharafriends/src/main/java/com/edw88/baricharafriends/ui/drawer/ui/gallery/GalleryViewModel.kt
@@ -1,13 +1,13 @@
-package com.edw88.baricharafriends.drawer.ui.slideshow
+package com.edw88.baricharafriends.ui.drawer.ui.gallery
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class GalleryViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
