+++ baricharafriends/src/main/java/com/edw88/baricharafriends/ui/drawer/ui/slideshow/SlideshowFragment.kt
@@ -1,4 +1,4 @@
-package com.edw88.baricharafriends.drawer.ui.home
+package com.edw88.baricharafriends.ui.drawer.ui.slideshow
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,13 +8,13 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.edw88.baricharafriends.databinding.FragmentHomeBinding
+import com.edw88.baricharafriends.databinding.FragmentSlideshowBinding
 
 
-class HomeFragment : Fragment() {
+class  SlideshowFragment : Fragment() {
 
-    private lateinit var homeViewModel: HomeViewModel
-    private var _binding: FragmentHomeBinding? = null
+    private lateinit var slideshowViewModel: SlideshowViewModel
+    private var _binding: FragmentSlideshowBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -25,14 +25,14 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        homeViewModel =
-            ViewModelProvider(this).get(HomeViewModel::class.java)
+        slideshowViewModel =
+            ViewModelProvider(this).get(SlideshowViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentSlideshowBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner, Observer {
+        val textView: TextView = binding.textSlideshow
+        slideshowViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
