+++ baricharafriends/src/main/java/com/edw88/baricharafriends/ui/drawer/ui/slideshow/SlideshowViewModel.kt
@@ -1,13 +1,13 @@
-package com.edw88.baricharafriends.drawer.ui.gallery
+package com.edw88.baricharafriends.ui.drawer.ui.slideshow
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class SlideshowViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
