+++ baricharafriends/src/main/java/com/edw88/baricharafriends/ui/list/PoiAdapter.kt
@@ -1,9 +1,8 @@
-package com.edw88.baricharafriends.list
+package com.edw88.baricharafriends.ui.list
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AdapterView
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
