+++ baricharafriends/src/main/java/com/edw88/baricharafriends/ui/main/MainActivity.kt
