+++ baricharafriends/src/main/res/layout/app_bar_drawer.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".drawer.DrawerActivity">
+    tools:context=".ui.drawer.DrawerActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
