+++ baricharafriends/src/main/res/layout/fragment_detail.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:backgroundTint="@color/cardview_shadow_start_color"
     android:backgroundTintMode="src_in"
-    tools:context="detail.DetailFragment">
+    tools:context=".ui.detail.DetailFragment">
 
     <TextView
         android:id="@+id/cementerio_text_view"
