+++ baricharafriends/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_list"
-        android:name="com.edw88.baricharafriends.list.ListFragment"
+        android:name="com.edw88.baricharafriends.ui.list.ListFragment"
         android:label="@string/title_list"
         tools:layout="@layout/fragment_list">
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/navigation_detail"
-        android:name="com.edw88.baricharafriends.detail.DetailFragment"
+        android:name="com.edw88.baricharafriends.ui.detail.DetailFragment"
         android:label="@string/title_detail"
         tools:layout="@layout/fragment_detail"
         >
