+++ baricharafriends/src/main/res/navigation/mobile_navigation2.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.edw88.baricharafriends.drawer.ui.home.HomeFragment"
+        android:name="com.edw88.baricharafriends.ui.drawer.ui.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.edw88.baricharafriends.drawer.ui.gallery.GalleryFragment"
+        android:name="com.edw88.baricharafriends.ui.drawer.ui.gallery.GalleryFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.edw88.baricharafriends.drawer.ui.slideshow.SlideshowFragment"
+        android:name="com.edw88.baricharafriends.ui.drawer.ui.slideshow.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 </navigation>
